Revert back to previous solution with non-lifted state and 'key' variable instead.
